% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anos_osf.R
\docType{data}
\name{anos_osf}
\alias{anos_osf}
\title{Identificadores (ids) no osf dos dados anuais do Catálogo de Teses e Dissertações da Fundação Coordenação de Aperfeiçoamento de Pessoal de Nível Superior (CAPES)}
\format{
Um data frame com as colunas:
\describe{
  \item{ano}{Ano dos dados (1987-2022).}
  \item{osf_id}{ID no OSF correspondente ao ano.}
}
}
\source{
\url{https://osf.io/}
}
\usage{
anos_osf
}
\description{
Um data frame contendo os anos e os ids correspondentes para baixar os arquivos.
}
\examples{
data(anos_osf)
head(anos_osf)
}
\keyword{datasets}
