% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_dados_capes.R
\name{baixar_dados_capes}
\alias{baixar_dados_capes}
\title{Baixar dados anuais do Catálogo de Teses e Dissertações da Fundação Coordenação de Aperfeiçoamento de Pessoal de Nível Superior (CAPES) de acordo com os anos selecionados. Os arquivos ficam disponíveis no diretório informado, em formato parquet.}
\usage{
baixar_dados_capes(anos, destino = tempdir())
}
\arguments{
\item{anos}{vetor com os anos desejados}

\item{destino}{diretório onde os arquivos serão salvos (padrão: diretório temporário)}
}
\value{
lista de caminhos dos arquivos baixados ou já existentes
}
\description{
Baixar dados anuais do Catálogo de Teses e Dissertações da Fundação Coordenação de Aperfeiçoamento de Pessoal de Nível Superior (CAPES) de acordo com os anos selecionados. Os arquivos ficam disponíveis no diretório informado, em formato parquet.
}
