% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_texto_capes.R
\name{buscar_texto_capes}
\alias{buscar_texto_capes}
\title{Buscar termos nos campos textuais dos dados Catálogo de Teses e Dissertações da Fundação Coordenação de Aperfeiçoamento de Pessoal de Nível Superior (CAPES)}
\usage{
buscar_texto_capes(termo, campo, anos, destino = tempdir())
}
\arguments{
\item{termo}{string, termo a ser buscado.}

\item{campo}{string, nome do campo onde buscar (ex.: "resumo", "titulo").}

\item{anos}{vetor de anos a serem considerados na busca.}

\item{destino}{diretório onde os arquivos estão ou serão salvos (padrão: `tempdir()`).}
}
\value{
um `data.frame` com as linhas que correspondem à busca.
}
\description{
Esta função permite buscar termos específicos nos campos textuais do Catálogo de Teses e Dissertações da CAPES.
Caso os arquivos ainda não estejam disponíveis no diretório especificado, eles serão baixados automaticamente.
}
