% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capes.R
\docType{data}
\name{capes_synthetic_df}
\alias{capes_synthetic_df}
\title{Synthetic CAPES Data}
\format{
A data frame with the following columns:
\describe{
  \item{base_year}{Reference year of the data.}
  \item{institution}{Higher Education Institution.}
  \item{area}{Area of Concentration.}
  \item{program_name}{Name of the Graduate Program.}
  \item{type}{Type of work (e.g., Master's, Doctorate).}
  \item{region}{Region of Brazil.}
  \item{state}{Federative Unit (state).}
  \item{n}{Total number of works.}
}
}
\source{
Synthetic data created from the CAPES Catalog of Theses and Dissertations.
}
\usage{
capes_synthetic_df
}
\description{
Aggregated data from the CAPES Catalog of Theses and Dissertations, 
containing summarized information by year, institution, area, program, type, region, and state (UF).
}
\examples{
data(capes_synthetic_df)
head(capes_synthetic_df)

}
\keyword{datasets}
