% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capes.R
\docType{data}
\name{capes_years}
\alias{capes_years}
\title{Annual files of the CAPES Catalog of Theses and Dissertations}
\format{
A data frame with one row per year (1987-2024) and the columns:
\describe{
  \item{year}{Year of the data.}
  \item{file}{File name (e.g. `capes_1987.parquet`).}
  \item{url}{Download URL of the file.}
  \item{bytes}{File size in bytes, used to check the download.}
  \item{md5}{MD5 checksum of the file.}
}
}
\source{
Derived from the CAPES open data portal,
  <https://dadosabertos.capes.gov.br/group/catalogo-de-teses-e-dissertacoes-brasil>.
}
\usage{
capes_years
}
\description{
A data frame listing, for each year, the Parquet file distributed by the
package and the URL from which it is downloaded. Used internally by
[download_capes_data()]; the host can be overridden with the
`capesR.base_url` option (see [download_capes_data()]).
}
\examples{
data(capes_years)
head(capes_years)

}
\keyword{datasets}
