% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_capes.R
\name{ler_dados_capes}
\alias{ler_dados_capes}
\title{Ler arquivos Parquet baixados e combinar os dados}
\usage{
ler_dados_capes(arquivos)
}
\arguments{
\item{arquivos}{Lista de caminhos dos arquivos Parquet}
}
\value{
Data frame combinado com todos os dados lidos
}
\description{
Ler arquivos Parquet baixados e combinar os dados
}
