% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capes.R
\docType{data}
\name{years_osf}
\alias{years_osf}
\title{Identifiers (IDs) on OSF for the annual data of the Catalog of Theses and Dissertations from the Brazilian Coordination for the Improvement of Higher Education Personnel (CAPES)}
\format{
A data frame with the following columns:
\describe{
  \item{year}{Year of the data (1987-2022).}
  \item{osf_id}{OSF ID corresponding to the year.}
}
}
\source{
\url{https://osf.io/}
}
\usage{
years_osf
}
\description{
A data frame containing the years and the corresponding IDs for downloading the files.
}
\examples{
data(years_osf)
head(years_osf)

}
\keyword{datasets}
