% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_texto_capes.R
\name{buscar_texto_capes}
\alias{buscar_texto_capes}
\title{Buscar termos nos campos textuais dos dados do Catálogo de Teses e Dissertações da CAPES}
\usage{
buscar_texto_capes(dados, termo, campo)
}
\arguments{
\item{dados}{`data.frame` contendo os dados do Catálogo de Teses e Dissertações da CAPES.}

\item{termo}{string, termo a ser buscado.}

\item{campo}{string, nome do campo onde buscar (ex.: "resumo", "titulo").}
}
\value{
um `data.frame` com as linhas que correspondem à busca ou uma mensagem informando que nenhum resultado foi encontrado.
}
\description{
Esta função permite buscar termos específicos nos campos textuais de um `data.frame` previamente carregado.
}
