% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capes.R
\name{download_capes_data}
\alias{download_capes_data}
\alias{baixar_dados_capes}
\title{Download CAPES Data}
\usage{
download_capes_data(years, destination = tempdir())

baixar_dados_capes(years, destination = tempdir())
}
\arguments{
\item{years}{A vector with the desired years.}

\item{destination}{The directory where the files will be saved (default: temporary directory).}
}
\value{
A list of file paths for the downloaded or already existing files.
}
\description{
Downloads CAPES theses and dissertations data files from OSF for selected years.
}
\examples{
\dontrun{
# Download data for the years 1987 and 1990
capes_files <- download_capes_data(c(1987, 1990))
}
}
