% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_capes.R
\name{ler_dados_capes}
\alias{ler_dados_capes}
\title{Ler e filtrar dados do Catálogo de Teses e Dissertações da CAPES}
\usage{
ler_dados_capes(arquivos, filtros = list())
}
\arguments{
\item{arquivos}{vetor ou lista de caminhos dos arquivos Parquet.}

\item{filtros}{lista de filtros para aplicar (ex.: list(ano_base = 1987, uf = "SP", titulo = "educação")).}
}
\value{
um `data.frame` contendo os dados combinados e filtrados.
}
\description{
Esta função combina os dados de múltiplos arquivos Parquet e aplica filtros opcionais, incluindo buscas textuais.
}
